% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Anglesea}
\alias{Anglesea}
\title{Example 2 from Hedges, Pustejovsky, & Shadish (2012)}
\format{
A data frame with 55 rows and 5 variables
}
\source{
Anglesea, M. M., Hoch, H., & Taylor, B. A. (2008). 
Reducing rapid eating in teenagers with autism: Use of a pager prompt. 
\emph{Journal of Applied Behavior Analysis, 41}(1), 107-111. 
doi:\href{http://doi.org/10.1901/jaba.2008.41-107}{10.1901/jaba.2008.41-107}
}
\description{
Data from an ABAB design conducted by Anglesea, Hoch, & Taylor (2008). 
The variables are as follows:
\itemize{
  \item \code{case} Case identifier.  
  \item \code{condition} Factor indicating baseline or treatment condition
  \item \code{phase} Study phase (including both control and treatment condition)
  \item \code{session} Measurement occasion
  \item \code{outcome} Total seconds of eating time
}
}
\references{
Hedges, L. V., Pustejovsky, J. E., & Shadish, W. R. (2012).
A standardized mean difference effect size for single case designs. 
\emph{Research Synthesis Methods, 3}, 224-239. doi:\href{http://doi.org/10.1002/jrsm.1052}{10.1002/jrsm.1052}
}
\keyword{datasets}
