% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scdhlm.R
\name{ARPsimulator}
\alias{ARPsimulator}
\title{A shiny interface for the scdhlm package}
\usage{
ARPsimulator()
}
\description{
An interactive shiny interface for estimating design-comparable
standardized mean difference effect sizes from single-case designs. Estimation methods
for multiple baseline and treatment reversal designs are available.
}

