% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{BartonArwood}
\alias{BartonArwood}
\title{Barton-Arwood, Wehby, & Falk (2005)}
\format{A data frame with 143 rows and 4 variables}
\source{
Barton-Arwood, S. M., Wehby, J. H., & Falk, K. B. (2005). Reading
  instruction for elementary-age students with emotional and behavioral
  disorders: Academic and behavioral outcomes. \emph{Exceptional Children, 72}(1),
  7-27. doi:10.1177/001440290507200101
}
\description{
Data from a multiple baseline design conducted by Barton-Arwood, Wehby, and
Falk (2005). The variables are as follows: \itemize{ \item \code{case}
Participant identifier \item \code{condition} Factor identifying
the phase of the design (A or B) \item \code{session} Measurement occasion 
\item \code{outcome} Oral reading fluency score (words per minute) }
}
\keyword{datasets}
