% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Bryant2018}
\alias{Bryant2018}
\title{Bryant et al. (2018)}
\format{
A data frame with 536 rows and 8 variables
}
\source{
Bryant, D. R., Bryant, B. R., Sorelle-Miner, D. A., Falcomata, T. S.
& Nozari, M. (2018).
Tier 3 intensified intervention for second grade students with severe
mathematics difficulties. \emph{Archives of Psychology, 2}(11), 1-24.
\doi{10.31296/aop.v2i11.86}
}
\description{
Data from a clustered multiple baseline across participants design conducted by
Bryant et al. (2018). The variables are as follows:
\itemize{
  \item \code{Study_ID}. Study identifier.
  \item \code{school}. School identifier.
  \item \code{group}. Group identifier.
  \item \code{case}. Student identifier.
  \item \code{treatment}. Indicator for treatment phase.
  \item \code{session}. Measurement occasion.
  \item \code{session_trt}. Measurement occasion times treatment phase.
  \item \code{outcome}. Texas Early Mathematics Inventory (TEMI-Aim Check) scores.
  \item \code{session_c}. Measurement occasion centered at the follow-up time.
}
}
\keyword{datasets}
