% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_BCSMD.R
\name{calc_BCSMD}
\alias{calc_BCSMD}
\title{A wrapper function for calculating design comparable effect sizes}
\usage{
calc_BCSMD(
  design,
  case,
  phase,
  session,
  outcome,
  cluster = NULL,
  series = NULL,
  center = 0,
  round_session = TRUE,
  treatment_name = NULL,
  FE_base = 0,
  RE_base = 0,
  RE_base_2 = NULL,
  FE_trt = 0,
  RE_trt = NULL,
  RE_trt_2 = NULL,
  corStruct = "AR1",
  varStruct = "hom",
  A = NULL,
  B = NULL,
  cover = 95,
  bound = 35,
  symmetric = TRUE,
  summary = TRUE,
  data = NULL,
  ...
)
}
\arguments{
\item{design}{Character string to specify whether data comes from a treatment
reversal (\code{"TR"}), multiple baseline across participants
(\code{"MBP"}), replicated multiple baseline across behaviors
(\code{"RMBB"}), or clustered multiple baseline across participants (\code{"CMB"}).}

\item{case}{vector of case indicators or name of a character or factor vector
within \code{data} indicating unique cases.}

\item{phase}{vector of treatment indicators or name of a character or factor
vector within \code{data} indicating unique treatment phases.}

\item{session}{vector of measurement occasions or name of numeric vector
within \code{data} of measurement times.}

\item{outcome}{vector of outcome data or name of numeric vector of outcome
data within \code{data}.}

\item{cluster}{(Optional) vector of cluster indicators or name of a character
or factor vector within \code{data} indicating clusters.}

\item{series}{(Optional) vector of series indicators or name of a character
or factor vector within \code{data} indicating series.}

\item{center}{Numeric value for the centering value for session. Default is 0.}

\item{round_session}{Logical indicating whether to round \code{session} to the nearest integer. Defaults to \code{TRUE}.}

\item{treatment_name}{(Optional) character string corresponding to the name
of the treatment phase.}

\item{FE_base}{numerical value indicating the fixed effect at the baseline
phase. Take value from 0 (the default) up to 6. The value of 0 represents
including level as a fixed effect. The value of 1 to 6 represents including
linear trend to sextic trend as a fixed effect.}

\item{RE_base}{numerical value indicating the random effect at the baseline
phase. Take value from 0 (the default) up to 6. The value of 0 represents
including level as a random effect. The value of 1 to 6 represents
including linear trend to sextic trend as a random effect.}

\item{RE_base_2}{numerical value indicating the random effect at the baseline
phase for the cluster level in clustered multiple baseline design across
participants or for the case level in replicated multiple baseline across
behaviors. Take value from 0 (the default) up to 6. The value of 0
represents including level as a random effect. The value of 1 to 6
represents including linear trend to sextic trend as a random effect.}

\item{FE_trt}{numerical value indicating the fixed effect at the treatment
phase. Take value from 0 (the default) up to 6. The value of 0 represents
including change in level as a fixed effect. The value of 1 to 6 represents
including change in linear trend to sextic trend as a fixed effect.}

\item{RE_trt}{numerical value indicating the random effect at the treatment
phase. Take value from 0 (the default) up to 6. The value of 0 represents
including change in level as a random effect. The value of 1 to 6
represents including change in linear trend to sextic trend as a random
effect.}

\item{RE_trt_2}{numerical value indicating the random effect at the treatment
phase for the cluster level in clustered multiple baseline design across
participants or for the case level in replicated multiple baseline across
behaviors. Take value from 0 (the default) up to 6. The value of 0
represents including change in level as a random effect. The value of 1 to
6 represents including change in linear trend to sextic trend as a random
effect.}

\item{corStruct}{(Optional) character string indicating the correlation
structure of session-level errors. Include \code{"AR1"} (default),
\code{"MA1"}, and \code{"IID"}.}

\item{varStruct}{(Optional) character string indicating the
heteroscedasticity structure of session-level errors. Include \code{"hom"}
(default) and \code{"het"}.}

\item{A}{the time point immediately before the start of treatment.}

\item{B}{the time point at which outcomes are measured in the hypothetical
between-group design.}

\item{cover}{confidence level.}

\item{bound}{numerical tolerance for non-centrality parameter in
\code{\link[stats]{qt}}.}

\item{symmetric}{If \code{TRUE} (the default), use a symmetric confidence
interval. If \code{FALSE}, use a non-central t approximation to obtain an
asymmetric confidence interval.}

\item{summary}{Logical indicating whether to return a data frame with effect
size estimates and other information. If \code{TRUE} (default), return a
dataframe containing the effect size estimate, standard error, confidence
interval, and other information. If \code{FALSE}, return a list with effect
size estimate, degrees of freedom, and other information.}

\item{data}{(Optional) dataset to use for analysis. Must be a
\code{data.frame}.}

\item{...}{further arguments.}
}
\value{
A data frame containing the design-comparable effect size estimate,
  standard error, confidence interval, and other information, or a list of
  `g_mlm()` object.
}
\description{
Clean single case design data for treatment reversal and
  multiple baseline designs. Fit a multi-level model using restricted maximum
  likelihood estimation. Estimate a standardized mean difference effect size.
}
\examples{
data(Laski)
calc_BCSMD(design = "MBP",
           case = case, phase = treatment,
           session = time, outcome = outcome,
           center = 4,
           FE_base = 0, RE_base = 0, FE_trt = 0,
           data = Laski)

data(Anglesea)
calc_BCSMD(design = "TR",
           case = case, phase = condition,
           session = session, outcome = outcome,
           treatment_name = "treatment",
           FE_base = 0, RE_base = 0, FE_trt = 0,
           data = Anglesea)

data(Thiemann2001)
calc_BCSMD(design = "RMBB",
           case = case, series = series, phase = treatment,
           session = time, outcome = outcome,
           FE_base = 0, RE_base = 0, RE_base_2 = 0, FE_trt = 0,
           data = Thiemann2001)

data(Bryant2018)
calc_BCSMD(design = "CMB",
           cluster = group, case = case, phase = treatment,
           session = session, outcome = outcome,
           treatment_name = "treatment",
           FE_base = 0, RE_base = 0, RE_base_2 = 0, FE_trt = 0,
           data = Bryant2018)


}
