% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_BCSMD.R
\name{calc_consts}
\alias{calc_consts}
\title{Calculate constants}
\usage{
calc_consts(
  method,
  design,
  center = 0,
  FE_base = 0,
  RE_base = 0,
  RE_base_2 = NULL,
  FE_trt = 0,
  RE_trt = NULL,
  RE_trt_2 = NULL,
  corStruct = "AR1",
  varStruct = "hom",
  A,
  B
)
}
\arguments{
\item{method}{Character string indicating the estimation method. Options
are \code{"RML"} or \code{"Bayes"}.}

\item{design}{Character string to specify whether data comes from a treatment
reversal (\code{"TR"}), multiple baseline across participants
(\code{"MBP"}), replicated multiple baseline across behaviors
(\code{"RMBB"}), or clustered multiple baseline across participants
(\code{"CMB"}).}

\item{center}{Numeric value for the centering value for session. Default is 0.}

\item{FE_base}{Vector of integers specifying which fixed effect terms to
include in the baseline phase. Setting \code{FE_base = 0} includes only a
level. Setting \code{FE_base = c(0,1)} includes a level and a linear time
trend.}

\item{RE_base}{Vector of integers specifying which random effect terms to
include in the baseline phase. Setting \code{RE_base = 0} includes only
levels (i.e., random intercepts). Setting \code{RE_base = c(0,1)} includes
random levels and random linear trends.}

\item{RE_base_2}{Vector of integers specifying which random effect terms to
include in the baseline phase for the cluster level in clustered multiple
baseline design across participants or for the case level in replicated
multiple baseline across behaviors. Setting \code{RE_base_2 = 0} includes
only levels (i.e., random intercepts). Setting \code{RE_base_2 = c(0,1)}
includes random levels and random linear trends.}

\item{FE_trt}{Vector of integers specifying which fixed effect terms to
include in the treatment phase. Setting \code{FE_trt = 0} includes only a
change in level. Setting \code{FE_trt = c(0,1)} includes a change in level
and a treatment-by-linear time trend.}

\item{RE_trt}{Vector of integers specifying which random effect terms to
include in the treatment phase. Setting \code{RE_trt = 0} includes only
random changes in level. Setting \code{RE_trt = c(0,1)} includes random
changes in level and random treatment-by-linear time trends.}

\item{RE_trt_2}{Vector of integers specifying which random effect terms to
include in the treatment phase for the cluster level in clustered multiple
baseline design across participants or for the case level in replicated
multiple baseline across behaviors. Setting \code{RE_trt_2 = 0} includes
only random changes in level. Setting \code{RE_trt_2 = c(0,1)} includes
random changes in level and random treatment-by-linear time trends.}

\item{corStruct}{(Optional) character string indicating the correlation
structure of session-level errors. Options are \code{"AR1"} (default),
\code{"MA1"}, or \code{"IID"}.}

\item{varStruct}{(Optional) character string indicating the
heteroscedasticity structure of session-level errors. Options are
\code{"hom"} (default) or \code{"het"}, which allows for the session-level
error variances to differ by phase.}

\item{A}{The time point immediately before the start of treatment in the
hypothetical between-group design.}

\item{B}{The time point at which outcomes are measured in the hypothetical
between-group design.}
}
\value{
A list of constants associated with the fixed effect parameters and
  the variance component parameters of the hierarchical model for calculating
  the effect size.
}
\description{
Calculate constants associated with the fixed effect parameters
  and the variance component parameters of the hierarchical model in order to
  calculate the BC-SMD effect size.
}
\examples{

calc_consts(method = "RML", design = "MBP", center = 4, 
            FE_base = c(0,1), RE_base = c(0),
            FE_trt = c(0,1), RE_trt = c(1),
            corStruct = "IID", varStruct = "hom",
            A = 4, B = 10) 


calc_consts(method = "Bayes", design = "MBP", center = 4,
            FE_base = c(0,1), RE_base = c(0,1), RE_base_2 = NULL,
            FE_trt = c(0,1), RE_trt = NULL, RE_trt_2 = NULL,
            corStruct = "AR1", varStruct = "hom",
            A = 4, B = 10)


}
