% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Carson}
\alias{Carson}
\title{Carson (2008)}
\format{
A data frame with 47 rows and 5 variables
}
\source{
Carson, K. D., Gast, D. L., & Ayres, K. M. (2008). Effects of a photo
  activity schedule book on independent task changes by students with
  intellectual disabilities in community and school job sites. \emph{European
  Journal of Special Needs Education, 23}, 269-279.
}
\description{
Data from a BAB design conducted by Carson, Gast, & Ayres (2008). The
variables are as follows: 
\itemize{ 
   \item \code{case} Participant identifier 
   \item \code{treatment} Factor describing the treatment condition
   \item \code{phase} Numeric describing the phase of the study design for each case 
   \item \code{outcome} Outcome scores 
   \item \code{time} Measurement occasion 
}
}
\keyword{datasets}
