% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Case}
\alias{Case}
\title{Case et al. (1992)}
\format{
A data frame with 56 rows and 4 variables
}
\source{
Case, L. P., Harris, K. R., & Graham, S. (1992). Improving the
  mathematical problem-solving skills of students with learning disabilities:
  Self-regulated strategy development. \emph{The journal of special education,
  26}(1), 1-1. \doi{10.1177/002246699202600101}
}
\description{
Data from a multiple baseline design conducted by Case, Harris, & Graham
(1992). The variables are as follows: \itemize{ \item \code{case}.
Participant identifier. \item \code{session}. Measurement occasion. \item
\code{condition}. Factor identifying the phase of the design (baseline or
treatment). \item \code{outcome}. Number of subtraction equations and answers
correct on each word problem probe. }
}
\keyword{datasets}
