% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Datchuk}
\alias{Datchuk}
\title{Datchuk (2016)}
\format{
A data frame with 74 rows and 4 variables
}
\source{
Datchuk, S. M. (2016). Writing simple sentences and descriptive
  paragraphs: Effects of an intervention on adolescents with writing
  difficulties. \emph{Journal of Behavioral Education, 25}(2), 166-188.
  \doi{10.1007/s10864-015-9236-x}
}
\description{
Data from a multiple baseline design conducted by Datchuk (2016). The
variables are as follows: \itemize{ \item \code{case}. Participant
identifier. \item \code{session}. Measurement occasion. \item
\code{condition}. Factor identifying the phase of the design (baseline or
treatment). \item \code{outcome}. Correct word sequences per minute. }
}
\keyword{datasets}
