% Generated by roxygen2 (4.0.1): do not edit by hand
\name{effect_size_MB}
\alias{effect_size_MB}
\title{Calculates HPS effect size}
\usage{
effect_size_MB(outcome, treatment, id, time, phi, rho)
}
\arguments{
\item{outcome}{Vector of outcome data. May not contain any missing values.}

\item{treatment}{Vector of treatment indicators. Must be the same length as \code{outcome}.}

\item{id}{factor vector indicating unique cases. Must be the same length as \code{outcome}.}

\item{time}{vector of measurement occasion times. Must be the same length as \code{outcome}.}

\item{phi}{Optional value of the auto-correlation nuisance parameter, to be used
in calculating the small-sample adjusted effect size}

\item{rho}{Optional value of the intra-class correlation nuisance parameter, to be used
in calculating the small-sample adjusted effect size}
}
\value{
A list with the following components
\tabular{ll}{
\code{g_dotdot} \tab total number of non-missing observations \cr
\code{K} \tab number of time-by-treatment groups containing at least one observation \cr
\code{D_bar} \tab numerator of effect size estimate \cr
\code{S_sq} \tab sample variance, pooled across time points and treatment groups \cr
\code{delta_hat_unadj} \tab unadjusted effect size estimate \cr
\code{phi} \tab corrected estimate of first-order auto-correlation \cr
\code{sigma_sq_w} \tab corrected estimate of within-case variance \cr
\code{rho} \tab estimated intra-class correlation \cr
\code{theta} \tab estimated scalar constant \cr
\code{nu} \tab estimated degrees of freedom \cr
\code{delta_hat} \tab corrected effect size estimate \cr
\code{V_delta_hat} \tab estimated variance of \code{delta_hat}
}
}
\description{
Calculates the HPS effect size estimator based on data from a multiple baseline design,
as described in Hedges, Pustejovsky, & Shadish (2013). Note that the data must contain one row per
measurement occasion per subject.
}
\note{
If phi or rho is left unspecified (or both), estimates for the nuisance
parameters will be calculated.
}
\examples{
data(Saddler)
with(subset(Saddler, measure==1), effect_size_MB(outcome, treatment, case, time))

data(Laski)
with(Laski, effect_size_MB(outcome, treatment, case, time))
}
\references{
Hedges, L. V., Pustejovsky, J. E., & Shadish, W. R. (2013).
A standardized mean difference effect size for multiple baseline designs.
\emph{Research Synthesis Methods}, forthcoming.
}

