% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REML-ES-functions.R
\name{extract_varcomp_ex}
\alias{extract_varcomp_ex}
\title{Extract estimated variance components}
\usage{
extract_varcomp_ex(m_fit)
}
\arguments{
\item{mod}{Fitted model of class lmeStruct.}
}
\value{
Object of class \code{varcomp} consisting of a list of estimated
  variance components: level-1 error variance, auto-correlation, and random
  effects variances.
}
\description{
Extracts the estimated variance components from a fitted linear
  mixed effects model (lmeStruct object).
}
\examples{

library(nlme)
Laski_RML <- lme(fixed = outcome ~ treatment,
                 random = ~ 1 | case,
                 correlation = corAR1(0, ~ time | case),
                 data = Laski)
extract_varcomp_ex(Laski_RML)

}
