% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_mlm_Bayes.R
\name{g_mlm_Bayes}
\alias{g_mlm_Bayes}
\title{Calculates a between-case standardized mean difference effect size estimate}
\usage{
g_mlm_Bayes(mod, p_const, r_const, rconst_base_var_index = 1, cover = 95)
}
\arguments{
\item{mod}{Fitted model of class brmsfit (estimated using
\code{brms::brm()}), from which to estimate the effect size.}

\item{p_const}{Vector of constants for calculating numerator of effect size.
Must be the same length as fixed effects in \code{mod}.}

\item{r_const}{Vector of constants for calculating denominator of effect
size. Must be the same length as the number of variance component
parameters in \code{mod_denom}.}

\item{rconst_base_var_index}{Something, not really sure what.}

\item{cover}{Confidence level.}
}
\value{
A list with the following components 
\tabular{ll}{ 
 \code{g_AB} \tab Posterior mean effect size estimate \cr
 \code{SE_g_AB} \tab Approximate standard error of mean effect size estimate \cr
 \code{nu} \tab Estimated denominator degrees of freedom \cr
 \code{CI_L} \tab Lower bound of credible interval for effect size \cr
 \code{CI_U} \tab Upper bound of credible interval for effect size \cr
 \code{es_num_vec} \tab Posterior samples of effect size numerator \cr
 \code{es_denom_vec} \tab Posterior samples of squared denominator of effect size \cr
 \code{autocor_param} \tab Posterior mean auto-correlation \cr
 \code{var_param} \tab Posterior mean of level-1 variance model parameter \cr
 \code{rho} \tab Posterior mean intra-class correlation \cr
}
}
\description{
Estimates a standardized mean difference effect size from a
  multi-level model estimated using \code{brms::brm}.
}
\references{
Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2014).
  Design-comparable effect sizes in multiple baseline designs: A general
  modeling framework. \emph{Journal of Educational and Behavioral Statistics,
  39}(4), 211-227. \doi{10.3102/1076998614547577}
}
