% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{GunningEspie}
\alias{GunningEspie}
\title{Gunning & Espie (2003)}
\format{
A data frame with 301 rows and 4 variables
}
\source{
Gunning, M. J., & Espie, C.A. (2003). Psychological treatment of
  reported sleep disorder in adults with intellectual disability using a
  multiple baseline design. \emph{Journal of Intellectual Disability
  Research, 47}(3), 191-202. \doi{10.1046/j.1365-2788.2003.00461.x}
}
\description{
Data from a multiple baseline design conducted by Gunning & Espie (2003). The
variables are as follows: 
\itemize{ 
  \item \code{case}. Participant identifier.  
  \item \code{session}. Measurement occasion. 
  \item \code{condition}. Factor identifying the phase of the design (baseline or treatment). 
  \item \code{outcome}. Sleep onset latency in minutes. 
}
}
\keyword{datasets}
