% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Lambert}
\alias{Lambert}
\title{Example 1 from Hedges, Pustejovsky, & Shadish (2012)}
\format{
A data frame with 461 rows and 6 variables
}
\source{
Lambert, M. C., Cartledge, G., Heward, W. L., & Lo, Y. (2006). 
Effects of response cards on disruptive behavior and academic responding 
during math lessons by fourth-grade urban students. 
\emph{Journal of Positive Behavior Interventions, 8}(2), 88-99. 
\doi{10.1177/10983007060080020701}
}
\description{
Data from an ABAB design conducted by Lambert, Cartledge, Heward, & Lo (2008). 
The variables are as follows:
\itemize{
  \item \code{measure}. Outcome measure description (disruptive behavior or academic response).
  \item \code{case}. Student identifier.  
  \item \code{treatment}. Factor indicating treatment or control condition. 
        SSR = single-subject responding. RC = response cards.
  \item \code{phase}. Study phase (including both control and treatment condition)
  \item \code{time}. Measurement occasion.
  \item \code{outcome}. Outcome scores.
}
}
\references{
Hedges, L. V., Pustejovsky, J. E., & Shadish, W. R. (2012).
A standardized mean difference effect size for single case designs. 
\emph{Research Synthesis Methods, 3}, 224-239. \doi{10.1002/jrsm.1052}
}
\keyword{datasets}
