% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmeInfo-imports.R
\name{lmeInfo}
\alias{lmeInfo}
\alias{extract_varcomp}
\alias{varcomp_vcov}
\alias{g_mlm}
\alias{CI_g}
\title{lmeInfo}
\description{
Functions imported from the lmeInfo package.

\itemize{
\item \code{\link[lmeInfo]{extract_varcomp}}
\item \code{\link[lmeInfo]{g_mlm}}
\item \code{\link[lmeInfo]{varcomp_vcov}}
\item \code{\link[lmeInfo]{CI_g}}
}
}
