\docType{data}
\name{MB1results}
\alias{MB1results}
\title{MB1 simulation results}
\format{A data frame}
\description{
Simulation results for model MB1 from Pustejovsky, Hedges,
& Shadish (2013).
}
\references{
Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2013).
Design-comparable effect sizes in multiple baseline
designs: A general approach to modeling and estimation.
}
\keyword{datasets}

