\docType{data}
\name{MB2results}
\alias{MB2results}
\title{MB2 simulation results}
\format{A data frame}
\description{
  Simulation results for model MB2 from Pustejovsky,
  Hedges, & Shadish (2013).
}
\references{
  Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R.
  (2013). Design-comparable effect sizes in multiple
  baseline designs: A general approach to modeling and
  estimation.
}
\keyword{datasets}

