% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Musser}
\alias{Musser}
\title{Musser (2001)}
\format{A data frame with 136 rows and 4 variables}
\source{
Musser, E. H., Bray, M. A., Kehle, T. J., & Jenson, W. R. (2001).
Reducing disruptive behaviors in students with serious emotional disturbance.
School Psychology Review, 30(2), 294-304.
}
\description{
Data from a multiple baseline design conducted by 
Musser, Bray, Kehle, & Jenson (2001). The variables are as follows:
\itemize{
  \item \code{student} Participant identifier
  \item \code{session} Measurement occasion
  \item \code{outcome} Percentage of disruptive intervals
  \item \code{treatment} Factor indicating baseline, treatment, or follow-up phase
}
}
\keyword{datasets}

