% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Peltier}
\alias{Peltier}
\title{Peltier et al. (2020)}
\format{
A data frame with 232 rows and 4 variables
}
\source{
Peltier, C., Sinclair, T. E., Pulos, J. M., & Suk, A. (2020). Effects
  of schema-based instruction on immediate, generalized, and combined
  structured word problems. \emph{The Journal of Special Education, 54}(2),
  101-112. \doi{10.1177/0022466919883397}
}
\description{
Data from a multiple baseline design conducted by Peltier, Sinclair, Pulos, &
Suk (2020). The variables are as follows: \itemize{ \item \code{case}.
Participant identifier. \item \code{session}. Measurement occasion. \item
\code{condition}. Factor identifying the phase of the design (baseline or
treatment). \item \code{outcome}. Mathematical problem-solving performance
(percentage). }
}
\keyword{datasets}
