% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-function.R
\name{phase_pairs}
\alias{phase_pairs}
\title{Calculate phase-pairs for a unique case}
\usage{
phase_pairs(x)
}
\arguments{
\item{x}{The subset of a single-case dataset for a unique case.}
}
\description{
Calculate phase-pairs based on phases and session numbering.
}
