% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-function.R
\name{preprocess_SCD}
\alias{preprocess_SCD}
\title{Clean Single Case Design Data}
\usage{
preprocess_SCD(
  data,
  case,
  phase,
  session,
  outcome,
  design,
  center = 0,
  treatment_name = NULL,
  round_session = TRUE
)
}
\arguments{
\item{data}{dataset to be cleaned. Must be a \code{data.frame}.}

\item{case}{name of a character or factor vector within \code{data} indicating unique cases.}

\item{phase}{name of a character or factor vector within \code{data} indicating unique treatment phases.}

\item{session}{name of numeric vector within \code{data} of measurement times.}

\item{outcome}{name of numeric vector of outcome data within \code{data}.}

\item{design}{Character string to specify whether data comes from a treatment reversal, "TR", or multiple baseline, "MB", design.}

\item{center}{Numeric value for the centering value for session. Default is 0.}

\item{treatment_name}{(Optional) Character string corresponding to the name of the treatment phase.}

\item{round_session}{Logical indicating whether to round \code{session} to the nearest integer. Defaults to \code{TRUE}.}
}
\value{
A cleaned SCD dataset that can be used for model fitting and effect size calculation.
}
\description{
Clean single case design data for treatment reversal and multiple baseline designs.
}
\note{
If treatment_name is left null it will choose the second level of the phase variable to be the treatment phase.
}
\examples{
data(Laski)
preprocess_SCD(data = Laski, case = case, 
               phase = treatment, session = time, 
               outcome = outcome, design = "MB", center = 4)

          
}
