% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Rodgers}
\alias{Rodgers}
\title{Rodgers et al. (2021)}
\format{
A data frame with 83 rows and 4 variables
}
\source{
Rodgers, D. B., Datchuk, S. M., & Rila, A. L. (2021).  Effects of a
  text-writing fluency intervention for postsecondary students with
  intellectual and developmental disabilities. \emph{Exceptionality, 29}(4),
  310-325. \doi{10.1080/09362835.2020.1850451}
}
\description{
Data from a multiple baseline design conducted by Rodgers, Datchuk, & Rila
(2021). The variables are as follows: \itemize{ \item \code{case}.
Participant identifier. \item \code{session}. Measurement occasion. \item
\code{condition}. Factor identifying the phase of the design (baseline or
treatment). \item \code{outcome}. The number of correct writing sequences in
1 minute. }
}
\keyword{datasets}
