% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Rodriguez}
\alias{Rodriguez}
\title{Rodriguez & Anderson (2014)}
\format{
A data frame with 148 rows and 4 variables
}
\source{
Rodriguez, B. J., & Anderson, C. M. (2014). Integrating a social
  behavior intervention during small group academic instruction using a total
  group criterion intervention. \emph{Journal of Positive Behavior Interventions,
  16}(4), 234-245. \doi{10.1177/1098300713492858}
}
\description{
Data from a multiple baseline design conducted by Rodriguez and Anderson (2014). 
The variables are as follows: \itemize{ 
  \item \code{case} Participant identifier 
  \item \code{condition} Factor identifying the phase of the design (A or B) 
  \item \code{session} Measurement occasion 
  \item \code{outcome} Percentage of intervals with problem behavior 
}
}
\keyword{datasets}
