% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Romaniuk}
\alias{Romaniuk}
\title{Romaniuk (2002)}
\format{
A data frame with 148 rows and 4 variables
}
\source{
Romaniuk, C., Miltenberger, R., Conyers, C., Jenner, N., Jurgens, M.,
  & Ringenberg, C. (2002). The influence of activity choice on problem
  behaviors maintained by escape versus attention. \emph{Journal of Applied
  Behavior Analysis, 35}(4), 349-62. 
  doi:\href{http://dx.doi.org/10.1901/jaba.2002.35-349}{10.1901/jaba.2002.35-349}
}
\description{
Data from a treatment reversal design conducted by Romaniuk and colleagues (2002).
The variables are as follows: \itemize{ 
  \item \code{case} Participant identifier 
  \item \code{phase} Factor identifying the phase of the design
  \item \code{condition} Factor identifying the treatment condition 
  \item \code{session} Measurement occasion 
  \item \code{outcome} Problem behavior
  \item \code{measurement} Character string describing how problem behavior was measured
}
}
\keyword{datasets}
