% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Ruiz}
\alias{Ruiz}
\title{Ruiz, et al. (2020)}
\format{A data frame with 840 rows and 5 variables}
\source{
Ruiz, F., Luciano, C., Florez, C., Suarez-Falcon, J., & Cardona-Betancourt, V. (2020).
  A Multiple-Baseline Evaluation of Acceptance and Commitment Therapy Focused on Repetitive Negative 
  Thinking for Comorbid Generalized Anxiety Disorder and Depression. Frontiers in Psychology, 11. 
  doi: \href{https://doi.org/10.3389/fpsyg.2020.00356}{10.3389/fpsyg.2020.00356}
}
\description{
Data from a multiple baseline design conducted by  Ruiz,
Luciano, Florez, Suarez-Falcon, &  Cardona-Betancourt (2020). The variables are as follows:
\itemize{
  \item \code{case}. Participant identifier.
  \item \code{measure}. Outcome measure description (AAQ-II, ANXIETY, CFQ, DASS-TOTAL, DEPRESSION, PSWQ, PTQ, STRESS, VQ-OBSTRUCTION, or VQ-PROGRESS).
  \item \code{treatment} Factor indicating baseline, treatment, post, or follow-up phase.
  \item \code{time}. Measurement occasion.
  \item \code{outcome}. Outcome scores.
}
}
\keyword{datasets}
