% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Saddler}
\alias{Saddler}
\title{Example 1 from Hedges, Pustejovsky, & Shadish (2013)}
\format{A data frame with 124 rows and 5 variables}
\source{
Saddler, B., Behforooz, B., & Asaro, K. (2008). 
The effects of sentence-combining instruction on the writing of 
fourth-grade students with writing difficulties. 
\emph{The Journal of Special Education, 42}(2), 79-90. 
doi:\href{http://doi.org/10.1177/0022466907310371}{10.1177/0022466907310371}
}
\description{
Data from a multiple baseline design conducted by 
Saddler, Behforooz, & Asaro, (2008). The variables are as follows:
\itemize{
  \item \code{case} Student identifier  
  \item \code{measure} Factor indicating the outcome measure (writing quality, T-unit length, number of constructions)
  \item \code{outcome} Value of outcome measure.
  \item \code{time}. Measurement occasion.
  \item \code{treatment}. Factor indicating the treatment phase.
}
}
\references{
Hedges, L. V., Pustejovsky, J. E., & Shadish, W. R. (2013). 
A standardized mean difference effect size for multiple baseline designs across individuals. 
\emph{Research Synthesis Methods, 4}(4), 324-341. doi:\href{http://doi.org/10.1002/jrsm.1086}{10.1002/jrsm.1086}
}
\keyword{datasets}
