% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Salazar}
\alias{Salazar}
\title{Salazar, et al. (2020)}
\format{
A data frame with 324 rows and 5 variables
}
\source{
Salazar, D., Ruiz, F., Ramírez, E., & Cardona-Betancourt, V. (2020).
  Acceptance and Commitment Therapy Focused on Repetitive Negative Thinking for
  Child Depression: A Randomized Multiple-Baseline Evaluation. The Psychological
  Record. \doi{10.1007/s40732-019-00362-5}
}
\description{
Data from a multiple baseline design conducted by Salazar, Ruiz,
Ramírez1, &  Cardona-Betancourt (2020). The variables are as follows:
\itemize{
  \item \code{case}. Participant identifier.
  \item \code{measure}. Outcome measure description (AFQ-Y, PTQ-C, or GPQ-C).
  \item \code{treatment} Factor indicating baseline, treatment, post, or follow-up phase.
  \item \code{time}. Measurement occasion.
  \item \code{outcome}. Outcome scores.
}
}
\keyword{datasets}
