% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REML-ES-functions.R
\name{scdhlm-deprecated}
\alias{scdhlm-deprecated}
\title{Deprecated functions in scdhlm}
\description{
This function still works but maybe removed in a later version of the package.
}
\details{
\itemize{
 \item \code{\g_REML}: This function is deprecated and may be removed in a later version of the package.
}
}
