\docType{data}
\name{Schutte}
\alias{Schutte}
\title{Example from Pustejovsky, Hedges, & Shadish (2013)}
\format{A data frame with 136 rows and 4 variables}
\source{
Schutte, N. S., Malouff, J. M., & Brown, R. F. (2008).
Efficacy of an emotion-focused treatment for prolonged
fatigue. \emph{Behavior Modification, 32}(5), 699-713. doi:
10.1177/0145445508317133
}
\description{
Data from a multiple baseline design conducted by Schutte,
Malouff, & Brown (2008). The variables are as follows:
\itemize{ \item \code{case}. Participant identifier.  \item
\code{week}. Measurement occasion.  \item \code{treatment}.
Factor indicating baseline or treatment phase.  \item
\code{fatigue}. Fatigue severity scale scores. }
}
\references{
Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2013).
Design-comparable effect sizes in multiple baseline
designs: A general approach to modeling and estimation.
}
\keyword{datasets}

