% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-function.R
\name{session_by_treatment}
\alias{session_by_treatment}
\title{Calculate session-by-treatment interactions for a unique case}
\usage{
session_by_treatment(phase, session, trt_phase)
}
\arguments{
\item{phase}{vector of treatment indicators or a character or factor vector indicating unique treatment phases.}

\item{session}{numeric vector of measurement occasions.}

\item{trt_phase}{character string indicating the phase value corresponding to
the treatment condition.}
}
\description{
Calculate session-by-treatment interactions based on phases and
  session numbering.
}
