% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm.R
\name{shine_scd}
\alias{shine_scd}
\title{A shiny interface for the scdhlm package}
\usage{
shine_scd()
}
\description{
An interactive shiny interface for estimating design-comparable
standardized mean difference effect sizes from single-case designs. Estimation methods
for multiple baseline and treatment reversal designs are available.
}
