% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Thiemann2001}
\alias{Thiemann2001}
\title{Thiemann & Goldstein (2001)}
\format{
A data frame with 363 rows and 8 variables
}
\source{
Thiemann, K.S., & Goldstein, H. (2001).
Social stories, written text cues, and video feedback: effects on social communication
of children with Autism. \emph{Journal of Applied Behavior Analysis, 34}(4), 425-446.
\doi{10.1901/jaba.2001.34-425}
}
\description{
Data from a multiple baseline across behaviors design conducted by
Thiemann & Goldstein (2001). The variables are as follows:
\itemize{
  \item \code{Study_ID}. Study identifier.
  \item \code{case}. Student identifier.
  \item \code{series}. Series identifier.
  \item \code{outcome}. Frequency of coded social communication skills, as measured by
  a direct observation coding system with 15-second intervals recoding for the occurrence
  of any of the four social measures: contingent responses, securing attention, initiating
  comments, and initiating requests.
  \item \code{time}. Measurement occasion.
  \item \code{treatment}. Indicator for treatment phase.
  \item \code{trt_time}. Measurement occasion times treatment phase.
  \item \code{time_c}. Measurement occasion centered at the follow-up time.
}
}
\keyword{datasets}
