% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Thiemann2004}
\alias{Thiemann2004}
\title{Thiemann & Goldstein (2004)}
\format{
A data frame with 408 rows and 8 variables
}
\source{
Thiemann, K.S., & Goldstein, H. (2004).
Effects of peer training and written text cueing on social communication of
school-age children with pervasive developmental disorder. \emph{Journal of Speech Language 
and Hearing Research, 47}(1), 126-144. doi:\href{https://doi.org/10.1044/1092-4388(2004/012)}{10.1044/1092-4388(2004/012)}
}
\description{
Data from a multiple baseline across behaviors design conducted by
Thiemann & Goldstein (2004). The variables are as follows:
\itemize{
  \item \code{Study_ID}. Study identifier.
  \item \code{case}. Student identifier.
  \item \code{series}. Series identifier.
  \item \code{outcome}. Frequency of unprompted targeted social communication skills,
  as measured by a direct observation, paper and pencil coding system during the 10-minute
  social activity for each behavior for all sessions.
  \item \code{time}. Measurement occasion.
  \item \code{treatment}. Indicator for treatment phase.
  \item \code{trt_time}. Measurement occasion times treatment phase.
  \item \code{time_c}. Measurement occasion centered at the follow-up time.
}
}
\keyword{datasets}
