% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{Thorne}
\alias{Thorne}
\title{Thorne (2005)}
\format{
A data frame with 776 rows and 6 variables
}
\source{
Thorne, S., & Kamps, D. (2008). The effects of a group contingency intervention on academic engagement and problem 
behavior of at-risk students. \emph{Behavior Analysis in Practice, 1}(2), 12-18.
}
\description{
Data from an ABAB design conducted by Thorne and Kamps (2008). The variables are as follows:
\itemize{
  \item \code{case}. Participant identifier.  
  \item \code{measure}. Outcome measure description (academic engagement or inappropriate verbalizations). 
  \item \code{session}. Measurement occasion.
  \item \code{phase_id}. Factor descibing the phase of the study design for each case.
  \item \code{phase_indicator}. Indicator variable equal to 1 during intervention phases.
  \item \code{outcome}. Outcome scores
}
}
\keyword{datasets}
