% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{DelemereDounavi}
\alias{DelemereDounavi}
\title{Delemere & Dounavi (2018)}
\format{
A data frame with 187 rows and 5 variables
}
\source{
Delemere, E., & Dounavi, K. (2018). Parent-implemented bedtime fading
  and positive routines for children with autism spectrum disorders.
  \emph{Journal of Autism and Developmental Disorders, 48}(4), 1002-1019.
  \doi{10.1007/s10803-017-3398-4}
}
\description{
Data from a multiple baseline design conducted by Delemere & Dounavi (2018).
The variables are as follows: 
\itemize{ 
  \item \code{intervention}. Type of intervention received: bedtime fading or positive routines.
  \item \code{case}. Participant identifier. 
  \item \code{session}. Measurement occasion. 
  \item \code{condition}. Factor identifying the phase of the design (baseline or treatment). 
  \item \code{outcome}. Total sleep onset latency in minutes across nights. 
}
}
\keyword{datasets}
