% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing-functions.R
\name{graph_SCD}
\alias{graph_SCD}
\title{Graph Single Case Design Data}
\usage{
graph_SCD(
  case,
  phase,
  session,
  outcome,
  design,
  treatment_name = NULL,
  model_fit = NULL,
  data = NULL
)
}
\arguments{
\item{case}{factor vector indicating unique cases or name of variable within \code{data}. Must be the same length as \code{outcome}.}

\item{phase}{vector of treatment indicators or name of variable within \code{data}. Must be the same length as \code{outcome}.}

\item{session}{vector of measurement occasion times or name of variable within \code{data}. Must be the same length as \code{outcome}.}

\item{outcome}{vector of outcome data or name of variable within \code{data}. May not contain any missing values.}

\item{design}{Specify whether it is a treatment reversal, "TR", or multiple baseline, "MB", design}

\item{treatment_name}{(Optional) value of the name of the treatment phase}

\item{model_fit}{(Optional) lme fitted model that adds predicted values to graph}

\item{data}{(Optional) dataset to use for analysis. Must be data.frame.}
}
\value{
A ggplot graph
}
\description{
Graphs single case design data for treatment reversal and multiple baseline designs.
}
\note{
If treatment_name is left null it will choose the second level of the phase variable to be the treatment phase.
}
\examples{
data(Anglesea)
graph_SCD(case=case, phase=condition, 
          session=session, outcome=outcome, 
          design="TR", treatment_name = "treatment", 
          data=Anglesea)
          
data(BartonArwood)
graph_SCD(case=case, phase=condition, 
          session=session, outcome=outcome, 
          design="MB", treatment_name = "B",  
          data=BartonArwood)

}
