% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdhlm-package.r
\docType{data}
\name{MB1results}
\alias{MB1results}
\title{MB1 simulation results}
\format{
A data frame
}
\description{
Simulation results for model MB1 from  Pustejovsky, Hedges, & Shadish (2014).
}
\references{
Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2014). 
Design-comparable effect sizes in multiple baseline designs: A general modeling framework.
\emph{Journal of Educational and Behavioral Statistics, 39}(4), 211-227. doi:\href{https://doi.org/10.3102/1076998614547577}{10.3102/1076998614547577}
}
\keyword{datasets}
