\docType{data}
\name{MB4results}
\alias{MB4results}
\title{MB4 simulation results}
\format{A data frame}
\description{
Simulation results for model MB4 from Pustejovsky, Hedges,
& Shadish (2013).
}
\references{
Pustejovsky, J. E., Hedges, L. V., & Shadish, W. R. (2013).
Design-comparable effect sizes in multiple baseline
designs: A general approach to modeling and estimation.
}
\keyword{datasets}

