% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess-function.R
\name{preprocess_SCD}
\alias{preprocess_SCD}
\title{Clean Single Case Design Data}
\usage{
preprocess_SCD(
  data,
  case,
  phase,
  session,
  outcome,
  design = c("MB", "TR"),
  center,
  treatment_name = NULL,
  round_session = TRUE
)
}
\arguments{
\item{data}{dataset to be cleaned. Must be data.frame.}

\item{case}{factor vector indicating unique cases or name of variable within \code{data}.}

\item{phase}{vector of treatment indicators or name of variable within \code{data}.}

\item{session}{vector of measurement occasion times or name of variable within \code{data}.}

\item{outcome}{vector of outcome data or name of variable within \code{data}.}

\item{design}{Specify whether it is a treatment reversal, "TR", or multiple baseline, "MB", design}

\item{center}{Specify the centering value for session.}

\item{treatment_name}{(Optional) value of the name of the treatment phase}

\item{round_session}{Default is rounding session to nearest integer.}
}
\value{
A cleaned SCD dataset that can be used for model fitting and effect size calculation.
}
\description{
Clean single case design data for treatment reversal and multiple baseline designs.
}
\note{
If treatment_name is left null it will choose the second level of the phase variable to be the treatment phase.
}
\examples{
data(Laski)
preprocess_SCD(data = Laski, case = case, phase = treatment, session = time, outcome = outcome, design = "MB", center = 4)

          
}
