\docType{data}
\name{Saddler}
\alias{Saddler}
\title{Example 1 from Hedges, Pustejovsky, & Shadish (2013)}
\format{A data frame with 124 rows and 5 variables}
\source{
  Saddler, B., Behforooz, B., & Asaro, K. (2008). The
  effects of sentence-combining instruction on the writing
  of fourth-grade students with writing difficulties.
  \emph{The Journal of Special Education, 42}(2), 79-90.
  doi:10.1177/0022466907310371
}
\description{
  Data from a multiple baseline design conducted by
  Saddler, Behforooz, & Asaro, (2008). The variables are as
  follows: \itemize{ \item \code{case}. Student identifier.
  \item \code{measure}. Outcome measure identifier.  1 =
  writing quality.  2 = T-unit length.  3 = number of
  constructions.  \item \code{outcome}. Outcome measure.
  \item \code{time}. Measurement occasion.  \item
  \code{treatment}. Indicator for treatment phase. }
}
\references{
  Hedges, L. V., Pustejovsky, J. E., & Shadish, W. R.
  (2013). A standardized mean difference effect size for
  multiple baseline designs. Working paper, Evanston, IL.
}
\keyword{datasets}

