% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REML-ES-functions.R
\name{scdhlm-deprecated}
\alias{scdhlm-deprecated}
\title{Deprecated functions in scdhlm}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
 \item \code{\link{extract_varcomp}}: `scdhlm::extract_varcomp()` will be updated with the `lmeInfo::extract_varcomp()` in the next version.
 \item \code{\link{Info_Expected_lmeAR1}}: This function is deprecated, and will be removed in the next version of this package.
 \item \code{\link{g_REML}}: This function is deprecated, and will be removed in the next version of this package.
}
}
