% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing-functions.R
\name{session_by_treatment}
\alias{session_by_treatment}
\title{Calculate session-by-treatment interactions for a unique case}
\usage{
session_by_treatment(x, trt_phase)
}
\arguments{
\item{x}{The subset of a single-case dataset for a unique case.}

\item{trt_phase}{character string indicating the phase value corresponding to
the treatment condition.}
}
\description{
Calculate session-by-treatment interactions based on phases and
  session numbering.
}
